+++ libs/dot-components/src/lib/components/link/Link.styles.ts
@@ -1,0 +1,16 @@
+import { Link } from '@material-ui/core';
+import styled, { css } from 'styled-components';
+
+export const rootClassName = 'dot-link';
+
+export const StyledLink = styled(Link)`
+  ${({ theme }) => css`
+    &.${rootClassName} {
+      cursor: pointer;
+
+      .MuiTypography-body1 {
+        margin-bottom: 0px;
+      }
+    }
+  `}
+`;